Force purging of WP core options from Redis cache when they are updated
It was reported that updating options in WordPress did not always update
to the visible web page. When debugged we found out that the update_option()
function did update the database contents, but not via object-cache.php
what was in Redis cache.

Many users simply stopped using object-cache.php because of this.
This workaround fix should solve the issue. The ultimate cause is some
kind of race condition in WP core code which we cannot solve right now.