List similar cruft on separate lines
With changes to lib/cruftfiles-ajax.php filenames are stored in
as a string in an array instead of arrays in an array.
This also requires that modules/cruftfiles.php is changed so that
it compares strings directly. It also allows for a more flexible
solutions ie. deleting multiple files at once.
Closes #107