Bugfix: don't load canonical-domain-and-https when WP-CLI is loaded
It seems that in multisite environments some of the $_SERVER variables are
already set, which ultimately leads to WP CLI not working due to a redirect
from canonical-domain-and-https. An example with logging:

multisite-testi@multisite-testi_ff0593:/data/wordpress/htdocs$ wp
$_SERVER['HTTP_HOST']: multisite-testi.ams-test.seravo.com
$_SERVER['REQUEST_URI']: /
Redirect to https://multisite-testi.ams-test.seravo.com/

The solution is to check that WP CLI is not loaded before attempting to
load the module itself. Fixes #134.