Make whole log viewer row clickable by displaying as block
Previously, if the text was shorter than the table row, clicking next to the
text (not directly on it) didn't make the cursor change to pointer and the
row contents were not added to the preview box. Now clicking anywhere on the
row makes the text appear in the preview box.