Bugfix: make dynamically added logrows appear in preview box on click
By directly tracking the click event for the "logrow" class, new dynamically
added DOM content (log rows appended by scrolling) were not tracked for the
click event, so clicking them did not make them appear in the log row
preview area. This problem is solved by tracking the click event for the
parent via event bubbling, but applying the "logrow" class selector to the
second parameter of the $.on function.