Remove the purge cache spinning animation in case of a failed POST
E.g. locally, purging the cache with older boxes could lead to problems (at
least with in my case). The failed errors are logged into the console but
the user is left in an undesired state looking at the loading animation
when the request has already failed.

There could be some improved error handling in the future, but this is good
enough for now.