Install specific versions of PHPCS and WP Coding Standards in Travis
There have been now a couple of incidents in which the PHP code sniff has
been successful locally, but then failed in the remote Travis CI build to
to different versions of PHPCS and WP Coding Standards. Instead of cloning
the repos directly from GitHub, we should stick to specific releases and
update them periodically.