Add optional parameter 'handled_http_codes' to api.php
With this it's possible to let the calling function decide if it wants to
handle errors. For example the code 400 (Bad request) could be useful to
not be masked by WP_Error when dealing with zone API.

I have reasoned that this does should not affect the current usage of
the functions.

Maybe add regex to this later? It would be pretty cool.

Examples of usage of the functions:
get_site_data( '/example', [200, 400]);
update_site_data( $data, '/example', [200, 400]);