Add module to allow seravotest user to bypass extra authentication
The Seravo tests need to login to a site to see that things in the wp-admin
work as expected. This can however be prevented by captchas or external
authentication integrations. With this module it will be possible to
bypass any custom authentication if necessary.

The test system that uses this bypass needs to generate a key on the server
side first and thus already have prior server access.