Limit shadow cookie lifetime to 12 hours
Sometimes users don't see the shadow banner or otherwise get confused
when returning to a site where they previously were in the shadow.
It is better to expire the shadow cookie in 12 hours (overnight) so that
if users return to a site the next day and have forgotten they were in
a shadow testing the previous day, they will automatically end back
on the production site, and need to manually go back into the shadow again.

It might also be that the shadow has been deleted, and since the cookie
expires automatically, users will not be stuck (at least for days) in a
situation where they have a shadow cookie to a site that does not load
anymore.

When exiting shadows, ensure cookies are not only empty, but also expired
and thus completely removed.