Improve shadow domain support
This improves the support of custom domain names for shadows.
As shadows don't have access to domain info of production container,
it's not really posssible to have full bulletproof implementation.

The old system functions the following way:
- Clear seravo_shadow cookie
  -> If set on different domain, user is stuck

The new fallback system:
- Check if seravo_shadow cookie is set
  -> If so, clear it
- If not, check if seravo_production cookie is set
  -> If so, redirect to it
- If not, check if subdomain is used (Doesn't work with TLD on second level eg. co.uk)
  -> If so, parse subdomain out and redirect to it
- Redirect to 'DEFAULT_DOMAIN' root path and trust that it redirects to WP_HOME

Shadows with domains will not set the seravo_shadow cookie but they still can be
accessed throught it (this doesn't break anything / backwards compatible).

Easiest ways to get stuck now are:
- Set seravo_shadow cookie manually (if you know how to set it, you know how to clear it)
- All of the following at the same time (really unlikely):
  * Access without instance-switcher (seravo_production not set)
  * Use shadow without subdomain
  * Redirection to WP_HOME is broken on production site