Redesign shadows postbox
Changes the width of shadows postbox from 100% to default to support
multiple postboxes on the same page (upcoming centralization changes).

To be able to fit all the functionality in a smaller postbox,
dropdown is used to select the shadow (like in mails-page) instead of showing all of them
in a table.

New design also has better shadow domain support.