Bugfix: use array as third $deps parameter for wp_enqueue_script
From https://developer.wordpress.org/reference/functions/wp_enqueue_script/:
The third parameter $deps is "An array of registered script handles this
script depends on". This bug resulted in a "jQuery is not defined" JS error
in the frontend when there were no other scripts that alredy loaded jQuery.
The situation could be replicated for instance in freshly deployed sites
on the frontend when the user is not logged in.