Remove WP-palvelu.fi notification fetcher
We have not used this feature for years so it can be considered deprecated.
We might want to reintroduce it some day, but it needs to be re-written
to use the SWD API etc. Now it is just best to remove it as it will also
stop the various sites from bombarding wp-palvelu.fi on every wp-admin
dashboard page load.