Fix login history only working on one timezone
Sites not using UTC are currently only shown empty table
on security page login history postbox.

The parser code was anyways very difficult to read so I simplified it a bit.
Using regex is probably not the best way to make it more readable but it
now takes care of the whole parsing process and is very easy to understand.