Make calls to wp-optimize-images run much faster
Do so in two ways:

1) Pass 3 extra command line values the PHP process already knows to
   the Python wp-optimize-images command, so that it can avoid doing
   any 'wp option get' itself, and thus save 1-2 seconds per invocation.

  $ time wp-optimize-images 1568x2844.png
  ...
  real	0m1.959s

  $ time wp-optimize-images --enable --set-max-resolution-width=4000 \
    --set-max-resolution-height=4000 1568x2844.png
  ...
  real	0m0.204s


2) Redirect output and background command so that exec() returns to PHP
   and continues without waiting.

There is no need to synchronize the roll-out of this with the roll-out
of the new version of wp-optimize-images.

This is safe to roll-out at any time, since if the --enable or other
arguments are not yet available in the environment, they the command will
just fail but exec() always continues nevertheless, so there is no draw-
back compared to previous code that had exec() disabled anyway.