Make Database pie chart a bit easier to read
Make the Database Size chart found under the Database page a bit
larger especially on fullHD monitors since before it was really small.

This is a temporary measure to make the sizing even a bit better
until a better solution is found.

Sizing not made perfect since there is issues in the upstream of
chart.js. Relates: #380