Add GUI for speed test
Under Tools -> Site Status there is now block for Speed Test to run
tests to different urls within the wordpress site. The test runs multiple
curl commands from within the server to see the speed of generating the
html from php. The tests are run with both with and without cache enabled.

There is button in the wordpress admin bar to test the speed of current page
using the above mentioned GUI.