Enhance disk space postbox
Use apexcharts.js for disk space postbox, implement a chart for total used disk space and redesign the per-directory chart.

Per-directory chart is now a bar chart that shows the size of certain directories listed in `sitestatus-ajax.php`.
The following logic is also followed:

- any directory under `/data` is shown
- all directories with size less than 1MB is not shown
- size of Seravo backups is not counted in the total used space nor is the directory listed in the bar chart

The donut chart for total disk use provides a simple view on how much of available space is used and the bar chart
provides a quick way to see what parts of the site take a lot of space.