Use a bar chart as database chart (Closes: #397)
Change database chart library to Apex Charts.
Use a bar chart instead of a pie chart.

If there are many plugins, for istance, the number of database tables can be really big,
so there is now a wrapper that makes the list scrollable if the number becomes too large.

Database table names are also aligned as far left as Apex Charts allows because table names can be long and be cut off.
Because of possible long names and the scroll bar introduced in this commit, there are now Apex Charts tooltips that have
the same information as the bar themselves.