Do not limit the number of shown large dirs
Before only 5 of the largest directories were included in the chart. Remove this limit. Because now the number of directories can be really big, there is now a wrapper that makes the list scrollable if it becomes too large. The dir names can be really long, so they are now aligned so far to the left as possible.