Fix and improve disk chart
The total disk space donut chart only had number values on the tooltips. Since they are always in MB, add the MB sign after the values.

Fixed the logic of calculating the used disk space and showing certain directories. Now directories are shown with the following logic:
- certain directories are always fetched
- directories with more than 100MB are always fetched

But there are directories the sizes of which are not counted in total usage:
- /data/backups
- /data/log
- /data/slog

Also, content of /data/backups are never shown even if their sizes exceed the 100M threshold.

This means the total disk usage is the size of /data minus the three dirs listed above.