Check Seravo request ID is set before accessing
It was noticed that in development environment our `log_queries()`
helper caused PHP notices like

> PHP Notice: Undefined index: HTTP_X_SERAVO_REQUEST_ID

This is caused by that fact that this variable is usually only available
in Seravo production environments. As Seravo Plugin can be used also in
development envs, we should always check if variable has been defined
before trying to access it.

Closes: #462