Fix margins
Fixes some of the margins in the plugin so elements are no longer right next to each other. Also fixes some of the small issues with the css with different screen sizes.

Some of the margins didn't work properly with different screen sizes and caused some of the elements to appear misaligned. After the fixes the elements should look okay on most modern screen sizes.