Hook successful login action in two-factor plugin
It was noticed that when user has enabled two-factor plugin on their
site, and provides correct credentials and 2FA/MFA succeeds, the login
event didn't end up in our login logs.

This happens because two-factor plugin modifies the login flow and shows
additional forms for doing the MFA validation.

Missing such information in logs it's at least a user experience
problem, as our customers assume they see all relevant events in the
logs. It might also be security issue, if trust model requires such data
to be in this log, in addition to other possible audit logs.