Add workflow for linting PHP style
Adds workflow for running PHP_CodeSniffer. Adds
custom coding standard based on PSR12. The standard
also includes PHPCompatiblity standard for checking
compatibility with PHP 7.0 and newer.

More changes will be made to the standard as needed.
None of the code supports the coding standard yet so
temporarily it's used to validate the empty 'test.php'
only.