Split PHP lints in separate workflow jobs
Some development tools can't be and don't need to be run
with the latest and oldest PHP versions we support.

Moves PHPCS in to a seperate workflow job from syntax linting
so it can be ran with PHP 8.2 instead of PHP 7.0 and 8.2.

This has no effect on the test results, PHPCS will still
also test for compatibility with PHP 7.0.